+++ src/Bootstrap.php
@@ -8,11 +8,12 @@ use PressbooksMultiInstitution\Actions\AssignBookToInstitution;
 use PressbooksMultiInstitution\Actions\AssignUserToInstitution;
 use PressbooksMultiInstitution\Actions\InstitutionalManagerDashboard;
 use PressbooksMultiInstitution\Controllers\AssignBooksController;
-use PressbooksMultiInstitution\Actions\PermissionsManager;
-use PressbooksMultiInstitution\Controllers\InstitutionsController;
 use PressbooksMultiInstitution\Controllers\AssignUsersController;
+use PressbooksMultiInstitution\Controllers\InstitutionsController;
 use PressbooksMultiInstitution\Services\InstitutionStatsService;
-use PressbooksMultiInstitution\Support\BookList;
+use PressbooksMultiInstitution\Services\PermissionsManager;
+use PressbooksMultiInstitution\Views\BookList;
+use PressbooksMultiInstitution\Views\UserList;
 
 /**
  * Class Bootstrap
@@ -40,6 +41,7 @@ final class Bootstrap
         $this->loadTranslations();
 
         Container::getInstance()->singleton(BookList::class, fn () => new BookList(app('db')));
+        Container::getInstance()->singleton(UserList::class, fn () => new UserList(app('db')));
     }
 
     public function registerMenus(): void
@@ -112,7 +114,7 @@ final class Bootstrap
         add_action('pb_new_blog', fn () => app(AssignBookToInstitution::class)->handle());
         add_action('network_admin_menu', fn () => app(PermissionsManager::class)->handleMenus(), 1000);
         add_action('admin_menu', fn () => app(PermissionsManager::class)->handleMenus(), 1000);
-        add_action('init', fn () => app(PermissionsManager::class)->setupInstitutionalFilters());
+        add_action('init', fn () => app(PermissionsManager::class)->setupFilters());
         add_action(
             'pb_institutional_after_save',
             fn ($newManagers, $revokedManagers) => app(PermissionsManager::class)->afterSaveInstitution($newManagers, $revokedManagers),
